Removed legacy Changelog implementation elements
