+++ apps/devportal/src/pages/changelog/[product]/rss.xml.tsx
@@ -2,13 +2,13 @@ import { CreateFeed } from '@lib/changelog/changelog-feeds';
 import { ChangelogEntriesByProduct } from 'sc-changelog/changelog';
 import GetProducts from 'sc-changelog/products';
 import { Product } from 'sc-changelog/types/product';
-import { slugify } from 'sc-changelog/utils/stringUtils';
+import { getSlug, slugify } from 'sc-changelog/utils/stringUtils';
 
 // Default export to prevent next.js errors
 const FeedPage = () => null;
 
 export async function getServerSideProps(context: any) {
-  const product = context.params.product;
+  const product = getSlug(context.params.product);
   const preview = context.preview ? context.preview : null;
   const products = await GetProducts(preview).then((response: Product[]) => {
     return response;
